Cleaned up Code
- Cleaned up errors, style mismatches
- Changed Snackbar handling to use one global snackbar
